+++ src/PhpWord/Element/Footnote.php
@@ -39,7 +39,6 @@ class Footnote extends Container
     public function __construct($paragraphStyle = null)
     {
         $this->containerType = 'footnote';
-
         // Set paragraph style
         if (is_array($paragraphStyle)) {
             $this->paragraphStyle = new Paragraph();