+++ src/PhpWord/Media.php
@@ -21,7 +21,7 @@ class Media
      *
      * @var array
      */
-    private static $_sectionMedia = array(
+    private static $sectionMedia = array(
         'images'     => array(),
         'embeddings' => array(),
         'links'      => array()
@@ -32,35 +32,42 @@ class Media
      *
      * @var array
      */
-    private static $_headerMedia = array();
+    private static $headerMedia = array();
 
     /**
      * Footer Media Elements
      *
      * @var array
      */
-    private static $_footerMedia = array();
+    private static $footerMedia = array();
+
+    /**
+     * Media elements
+     *
+     * @var array
+     */
+    private static $media = array();
 
     /**
      * ObjectID Counter
      *
      * @var int
      */
-    private static $_objectId = 1325353440;
+    private static $objectId = 1325353440;
 
     /**
      * Add new Section Media Element
      *
      * @param  string $src
      * @param  string $type
-     * @param  \PhpOffice\PhpWord\Element\Image $image
+     * @param  Image $image
      * @return mixed
      */
     public static function addSectionMediaElement($src, $type, Image $image = null)
     {
         $mediaId = md5($src);
         $key = ($type === 'image') ? 'images' : 'embeddings';
-        if (!array_key_exists($mediaId, self::$_sectionMedia[$key])) {
+        if (!array_key_exists($mediaId, self::$sectionMedia[$key])) {
             $cImg = self::countSectionMediaElements('images');
             $cObj = self::countSectionMediaElements('embeddings');
             $rID = self::countSectionMediaElements() + 7;
@@ -90,17 +97,17 @@ class Media
             $media['target'] = "$folder/section_$file";
             $media['type'] = $type;
             $media['rID'] = $rID;
-            self::$_sectionMedia[$key][$mediaId] = $media;
+            self::$sectionMedia[$key][$mediaId] = $media;
             if ($type === 'oleObject') {
-                return array($rID, ++self::$_objectId);
+                return array($rID, ++self::$objectId);
             }
             return $rID;
         } else {
             if ($type === 'oleObject') {
-                $rID = self::$_sectionMedia[$key][$mediaId]['rID'];
-                return array($rID, ++self::$_objectId);
+                $rID = self::$sectionMedia[$key][$mediaId]['rID'];
+                return array($rID, ++self::$objectId);
             }
-            return self::$_sectionMedia[$key][$mediaId]['rID'];
+            return self::$sectionMedia[$key][$mediaId]['rID'];
         }
     }
 
@@ -119,7 +126,7 @@ class Media
         $link['rID'] = $rID;
         $link['type'] = 'hyperlink';
 
-        self::$_sectionMedia['links'][] = $link;
+        self::$sectionMedia['links'][] = $link;
 
         return $rID;
     }
@@ -133,12 +140,12 @@ class Media
     public static function getSectionMediaElements($key = null)
     {
         if (!is_null($key)) {
-            return self::$_sectionMedia[$key];
+            return self::$sectionMedia[$key];
         }
 
-        $arrImages = self::$_sectionMedia['images'];
-        $arrObjects = self::$_sectionMedia['embeddings'];
-        $arrLinks = self::$_sectionMedia['links'];
+        $arrImages = self::$sectionMedia['images'];
+        $arrObjects = self::$sectionMedia['embeddings'];
+        $arrLinks = self::$sectionMedia['links'];
         return array_merge($arrImages, $arrObjects, $arrLinks);
     }
 
@@ -151,12 +158,12 @@ class Media
     public static function countSectionMediaElements($key = null)
     {
         if (!is_null($key)) {
-            return count(self::$_sectionMedia[$key]);
+            return count(self::$sectionMedia[$key]);
         }
 
-        $cImages = count(self::$_sectionMedia['images']);
-        $cObjects = count(self::$_sectionMedia['embeddings']);
-        $cLinks = count(self::$_sectionMedia['links']);
+        $cImages = count(self::$sectionMedia['images']);
+        $cObjects = count(self::$sectionMedia['embeddings']);
+        $cLinks = count(self::$sectionMedia['links']);
         return ($cImages + $cObjects + $cLinks);
     }
 
@@ -165,41 +172,12 @@ class Media
      *
      * @param  int $headerCount
      * @param  string $src
-     * @param  \PhpOffice\PhpWord\Element\Image $image
+     * @param  Image $image
      * @return int
      */
     public static function addHeaderMediaElement($headerCount, $src, Image $image = null)
     {
-        $mediaId = md5($src);
-        $key = 'header' . $headerCount;
-        if (!array_key_exists($key, self::$_headerMedia)) {
-            self::$_headerMedia[$key] = array();
-        }
-        if (!array_key_exists($mediaId, self::$_headerMedia[$key])) {
-            $cImg = self::countHeaderMediaElements($key);
-            $rID = $cImg + 1;
-            $cImg++;
-            $media = array();
-            $isMemImage = false;
-            if (!is_null($image)) {
-                $isMemImage = $image->getIsMemImage();
-                $extension = $image->getImageExtension();
-            }
-            if ($isMemImage) {
-                $media['isMemImage'] = true;
-                $media['createfunction'] = $image->getImageCreateFunction();
-                $media['imagefunction'] = $image->getImageFunction();
-            }
-            $file = 'image' . $cImg . '.' . strtolower($extension);
-            $media['source'] = $src;
-            $media['target'] = 'media/' . $key . '_' . $file;
-            $media['type'] = 'image';
-            $media['rID'] = $rID;
-            self::$_headerMedia[$key][$mediaId] = $media;
-            return $rID;
-        } else {
-            return self::$_headerMedia[$key][$mediaId]['rID'];
-        }
+        return self::addMediaElement("header{$headerCount}", 'image', $src, $image);
     }
 
     /**
@@ -210,17 +188,26 @@ class Media
      */
     public static function countHeaderMediaElements($key)
     {
-        return count(self::$_headerMedia[$key]);
+        return self::countMediaElements($key);
     }
 
     /**
      * Get Header Media Elements
      *
-     * @return int
+     * @return array
      */
-    public static function getHeaderMediaElements()
+    public static function getHeaderMediaElements($prefix = 'header')
     {
-        return self::$_headerMedia;
+        $mediaCollection = array();
+        if (!empty(self::$media)) {
+            foreach (self::$media as $key => $val) {
+                if (substr($key, 0, 6) == $prefix) {
+                   $mediaCollection[$key] = $val;
+                }
+            }
+        }
+
+        return $mediaCollection;
     }
 
     /**
@@ -228,41 +215,12 @@ class Media
      *
      * @param  int $footerCount
      * @param  string $src
-     * @param  \PhpOffice\PhpWord\Element\Image $image
+     * @param  Image $image
      * @return int
      */
     public static function addFooterMediaElement($footerCount, $src, Image $image = null)
     {
-        $mediaId = md5($src);
-        $key = 'footer' . $footerCount;
-        if (!array_key_exists($key, self::$_footerMedia)) {
-            self::$_footerMedia[$key] = array();
-        }
-        if (!array_key_exists($mediaId, self::$_footerMedia[$key])) {
-            $cImg = self::countFooterMediaElements($key);
-            $rID = $cImg + 1;
-            $cImg++;
-            $media = array();
-            $isMemImage = false;
-            if (!is_null($image)) {
-                $isMemImage = $image->getIsMemImage();
-                $extension = $image->getImageExtension();
-            }
-            if ($isMemImage) {
-                $media['isMemImage'] = true;
-                $media['createfunction'] = $image->getImageCreateFunction();
-                $media['imagefunction'] = $image->getImageFunction();
-            }
-            $file = 'image' . $cImg . '.' . strtolower($extension);
-            $media['source'] = $src;
-            $media['target'] = 'media/' . $key . '_' . $file;
-            $media['type'] = 'image';
-            $media['rID'] = $rID;
-            self::$_footerMedia[$key][$mediaId] = $media;
-            return $rID;
-        } else {
-            return self::$_footerMedia[$key][$mediaId]['rID'];
-        }
+        return self::addMediaElement("footer{$footerCount}", 'image', $src, $image);
     }
 
     /**
@@ -273,16 +231,120 @@ class Media
      */
     public static function countFooterMediaElements($key)
     {
-        return count(self::$_footerMedia[$key]);
+        return self::countMediaElements($key);
     }
 
     /**
      * Get Footer Media Elements
      *
-     * @return int
+     * @return array
      */
     public static function getFooterMediaElements()
     {
-        return self::$_footerMedia;
+        return self::getHeaderMediaElements('footer');
+    }
+
+    /**
+     * Add new media element
+     *
+     * @param string $container section|header|footer|footnotes
+     * @param string $mediaType image|embedding|hyperlink
+     * @param string $source
+     * @param Image $image
+     * @return int
+     */
+    public static function addMediaElement($container, $mediaType, $source, Image $image = null)
+    {
+        // Assign media Id and initiate media container if none exists
+        $mediaId = md5($source);
+        if (!array_key_exists($container, self::$media)) {
+            self::$media[$container]= array();
+        }
+
+        // Add media if not exists or point to existing media
+        if (!array_key_exists($mediaId, self::$media[$container])) {
+            $mediaCount = self::countMediaElements($container);
+            $mediaTypeCount = self::countMediaElements($container, $mediaType);
+            $mediaData = array();
+            $relId = $mediaCount + 1;
+            $mediaTypeCount++;
+            if ($mediaType == 'image') {
+                $isMemImage = false;
+                if (!is_null($image)) {
+                    $isMemImage = $image->getIsMemImage();
+                    $extension = $image->getImageExtension();
+                }
+                if ($isMemImage) {
+                    $mediaData['isMemImage'] = true;
+                    $mediaData['createfunction'] = $image->getImageCreateFunction();
+                    $mediaData['imagefunction'] = $image->getImageFunction();
+                }
+                $file = 'image' . $mediaTypeCount . '.' . strtolower($extension);
+                if ($container != 'footnotes') {
+                    $file = $container . '_' . $file;
+                }
+                $target = 'media/' . $file;
+            } elseif ($mediaType == 'hyperlink') {
+                $target = $source;
+            }
+            $mediaData['source'] = $source;
+            $mediaData['target'] = $target;
+            $mediaData['type'] = $mediaType;
+            $mediaData['rID'] = $relId;
+            self::$media[$container][$mediaId] = $mediaData;
+
+            return $relId;
+        } else {
+            return self::$media[$container][$mediaId]['rID'];
+        }
+    }
+
+    /**
+     * Get media elements count
+     *
+     * @param string $container
+     * @param string $mediaType
+     * @return int
+     */
+    public static function countMediaElements($container, $mediaType = null)
+    {
+        $mediaCount = 0;
+        foreach (self::$media[$container] as $mediaKey => $mediaData) {
+            if (!is_null($mediaType)) {
+                if ($mediaType == $mediaData['type']) {
+                    $mediaCount++;
+                }
+            } else {
+                $mediaCount++;
+            }
+        }
+
+        return $mediaCount;
+    }
+
+    /**
+     * Get media elements
+     *
+     * @param string $container
+     * @return int
+     */
+    public static function getMediaElements($container, $mediaType = null)
+    {
+        if (!array_key_exists($container, self::$media)) {
+            return false;
+        }
+
+        $mediaElements = array();
+        foreach (self::$media[$container] as $mediaKey => $mediaData) {
+            if (!is_null($mediaType)) {
+                if ($mediaType == $mediaData['type']) {
+                    $mediaElements[$mediaKey] = $mediaData;
+                }
+            } else {
+                $mediaElements[$mediaKey] = $mediaData;
+            }
+        }
+
+        return $mediaElements;
     }
 }