+++ src/PhpWord/Shared/XMLWriter.php
@@ -24,6 +24,8 @@ if (!defined('DATE_W3C')) {
  * @method bool startElement(string $name)
  * @method bool writeAttribute(string $name, string $value)
  * @method bool endElement()
+ * @method bool startDocument(string $version = 1.0, string $encoding = null, string $standalone = null)
+ * @method bool text(string $content)
  */
 class XMLWriter
 {