+++ src/PhpWord/Writer/Word2007/Footnotes.php
@@ -12,6 +12,7 @@ namespace PhpOffice\PhpWord\Writer\Word2007;
 use PhpOffice\PhpWord\Element\Footnote;
 use PhpOffice\PhpWord\Element\Text;
 use PhpOffice\PhpWord\Element\Link;
+use PhpOffice\PhpWord\Element\Image;
 use PhpOffice\PhpWord\Element\TextBreak;
 use PhpOffice\PhpWord\Style\Paragraph;
 use PhpOffice\PhpWord\Shared\XMLWriter;
@@ -34,17 +35,20 @@ class Footnotes extends Base
         // XML header
         $xmlWriter->startDocument('1.0', 'UTF-8', 'yes');
         $xmlWriter->startElement('w:footnotes');
-        $xmlWriter->writeAttribute(
-            'xmlns:r',
-            'http://schemas.openxmlformats.org/officeDocument/2006/relationships'
-        );
-        $xmlWriter->writeAttribute(
-            'xmlns:w',
-            'http://schemas.openxmlformats.org/wordprocessingml/2006/main'
-        );
+
+        $xmlWriter->writeAttribute('xmlns:ve', 'http://schemas.openxmlformats.org/markup-compatibility/2006');
+        $xmlWriter->writeAttribute('xmlns:o', 'urn:schemas-microsoft-com:office:office');
+        $xmlWriter->writeAttribute('xmlns:r', 'http://schemas.openxmlformats.org/officeDocument/2006/relationships');
+        $xmlWriter->writeAttribute('xmlns:m', 'http://schemas.openxmlformats.org/officeDocument/2006/math');
+        $xmlWriter->writeAttribute('xmlns:v', 'urn:schemas-microsoft-com:vml');
+        $xmlWriter->writeAttribute('xmlns:wp', 'http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing');
+        $xmlWriter->writeAttribute('xmlns:w10', 'urn:schemas-microsoft-com:office:word');
+        $xmlWriter->writeAttribute('xmlns:w', 'http://schemas.openxmlformats.org/wordprocessingml/2006/main');
+        $xmlWriter->writeAttribute('xmlns:wne', 'http://schemas.microsoft.com/office/word/2006/wordml');
+
         // Separator and continuation separator
         $xmlWriter->startElement('w:footnote');
-        $xmlWriter->writeAttribute('w:id', 0);
+        $xmlWriter->writeAttribute('w:id', -1);
         $xmlWriter->writeAttribute('w:type', 'separator');
         $xmlWriter->startElement('w:p');
         $xmlWriter->startElement('w:r');
@@ -53,9 +57,8 @@ class Footnotes extends Base
         $xmlWriter->endElement(); // w:r
         $xmlWriter->endElement(); // w:p
         $xmlWriter->endElement(); // w:footnote
-        // Content
         $xmlWriter->startElement('w:footnote');
-        $xmlWriter->writeAttribute('w:id', 1);
+        $xmlWriter->writeAttribute('w:id', 0);
         $xmlWriter->writeAttribute('w:type', 'continuationSeparator');
         $xmlWriter->startElement('w:p');
         $xmlWriter->startElement('w:r');
@@ -64,6 +67,7 @@ class Footnotes extends Base
         $xmlWriter->endElement(); // w:r
         $xmlWriter->endElement(); // w:p
         $xmlWriter->endElement(); // w:footnote
+        // Content
         foreach ($allFootnotesCollection as $footnote) {
             if ($footnote instanceof Footnote) {
                 $this->writeFootnote($xmlWriter, $footnote);
@@ -113,6 +117,8 @@ class Footnotes extends Base
                     $this->writeText($xmlWriter, $element, true);
                 } elseif ($element instanceof Link) {
                     $this->writeLink($xmlWriter, $element, true);
+                } elseif ($element instanceof Image) {
+                    $this->writeImage($xmlWriter, $element, true);
                 } elseif ($element instanceof TextBreak) {
                     $xmlWriter->writeElement('w:br');
                 }